fix: move dev intro IconsText component to dapps page
