+++ public/resources/en.json
@@ -3,12 +3,6 @@
   "landing_page": {
     "intro_heading": "Community <br />Governed <br />Blockchain",
     "intro_content": "<strong>Built on EOSIO.</strong> Delegated Proof of Stake is the future of smart-contract blockchains. We combine the blazing fast execution and scalability of EOSIO software with true decentralization that makes DPoS sustainable. Telos will have the most even money supply distribution of any blockchain. This token distribution improves governance so the network is more secure and ensures that the future of Telos is decided by its community. Telos has built the most advanced smart contract-controlled governance features of any blockchain.",
-    "developers_heading": "Developer Focus",
-    "developers_content": "Telos exists to provide a secure, stable, cost-effective network for deploying world class Dapps. Developers can expect affordable RAM prices and the ability to launch open-source or proprietary smart contracts with no transaction fees.",
-    "dpos_heading": "DPoS Speed",
-    "dpos_content": "The Telos network handles up to 5,000 transactions per second for blazing fast transactions and massive scalability. This means Dapps can reach a worldwide audience at the fastest average transaction speed of any blockchain—just 1.5 seconds!",
-    "gov_heading": "Good Governance",
-    "gov_content": "Telos has a rich governance system from voting amendments to the user agreement, to ongoing block producer elections, to worker proposal funding and dispute resolution. No other blockchain offers the cohesive governance of Telos.",
     "white_paper_heading": "Download the<br /> white paper",
     "white_paper_content": "Read the Telos white paper to judge for yourself.  We are out to create the best blockchain ever. Join us and build on Telos.",
     "github_text": "Visit our Github",
@@ -64,6 +58,12 @@
         "We are building a repository of sample contracts for new tokens, migration of existing tokens from Ethereum or EOS, voting functions, and other Telos development concepts. We also have a strong developer community to offer help in our Telegram groups, especially <a href=\"https://t.me/dappstelos\" target=\"_blank\" rel=\"noopener noreferrer\">Telos Dapp Development</a>.",
         "There is also an excellent tutorial series about deploying on Telos called <a href=\"https://link.medium.com/0UUfxmsMAV\" target=\"_blank\" rel=\"noopener noreferrer\">SensorDApp</a>."
       ],
+      "developers_heading": "Developer Focus",
+      "developers_content": "Telos exists to provide a secure, stable, cost-effective network for deploying world class Dapps. Developers can expect affordable RAM prices and the ability to launch open-source or proprietary smart contracts with no transaction fees.",
+      "dpos_heading": "DPoS Speed",
+      "dpos_content": "The Telos network handles up to 5,000 transactions per second for blazing fast transactions and massive scalability. This means Dapps can reach a worldwide audience at the fastest average transaction speed of any blockchain—just 1.5 seconds!",
+      "gov_heading": "Good Governance",
+      "gov_content": "Telos has a rich governance system from voting amendments to the user agreement, to ongoing block producer elections, to worker proposal funding and dispute resolution. No other blockchain offers the cohesive governance of Telos.",
       "tutorials_button": "TUTORIALS",
       "tutorials_description": "Learn how to build your Dapp on the Telos network.",
       "smart_contracts_button": "SMART CONTRACT TEMPLATES",
